workers/model_moondream/requirements.txt
numpy<2
torch==2.2.2
transformers==4.39.3
accelerate
pillow
redis
pydantic
